+++ bobot/Response.py
@@ -1,138 +1,49 @@
 """
-    User response module
+    User responses module.
+    Contains:
+        Message
+        Text
+        Keyboard
+        Location
+        File
+        Photo
+        Contact
 """
 
-from bobot.utils.utils import flatten
-from bobot.Errors import RuleNameError, ResponseFormatError, ResponseMessageError
+from bobot.utils.utils import instanceof, omit, pickCompat
+from bobot.Errors import ResponseFormatError, ResponseMessageError
 
-def createMessage(message, data):
-    """
-        Creates message from message description object
-        and update's data
-    """
-    options = {}
-    text = message
+class Message:
+    "Main Message class container"
 
-    if isinstance(message, dict):
-        text = message.get('text')
-        options = message.get('options', {})
+    method = 'sendMessage'
 
-        if not text:
-            raise ResponseMessageError('Message description should have "text" [String] property', 'text')
+    def __init__(self, **options):
+        self.options = options
 
-        if message.get('interpolate', False):
-            text = text.format(
-                text=data.get('text'),
-                first_name=data.get('first_name'),
-                username=data.get('username'),
-                second_name=data.get('second_name'))
+    def getData(self, data):
+        # pylint: disable=unused-argument
+        "Return data to send with method"
 
-    return text, options
+        return []
 
-def processResponseBody(responseBody):
-    "Processing response description body to actions lists"
+    def getOptions(self):
+        "Return options dict"
 
-    if isinstance(responseBody, dict):
-        return [responseBody]
-    elif isinstance(responseBody, list):
-        return list(map(processResponseBody, responseBody))
-    else:
-        raise ResponseFormatError('Wrong response description. Expected dict, got {}'.format(type(responseBody)))
+        options = self.options or {}
+        options.update({
+            'disable_notification': self.options.get('sielent', False),
+            'reply_to_message_id': self.options.get('replyId', None),
+        })
 
-def sendMessage(message=''):
-    "Returns message sending function"
+        if self.options.get('markup'):
+            options.update({'reply_markup': self.options.get('markup', None)})
 
-    def action(bot, chatId, data):
-        "Sends message"
-        text, options = createMessage(message, data)
-        return bot.sendMessage(chatId, text, options)
-    return action
+        return options
 
-def sendMessages(*messages):
-    "Returns messages sending function"
+    def run(self, bot, update):
+        "Apply update to response body"
 
-    def action(bot, chatId, data):
-        "Sends messages"
-        for message in messages:
-            text, options = createMessage(message, data)
-            bot.sendMessage(chatId, text, options)
-    return action
-
-def sendKeyboard(kb={}):
-    "Returns function that sends keyboard to user"
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.keyboard(chatId, kb)
-    return action
-
-def sendPhoto(photo=None, caption=''):
-    "Returns function that sends photo to user"
-    if not photo:
-        raise ResponseFormatError('Specify photo by URL or File')
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendPhoto(chatId, photo, caption)
-
-    return action
-
-def sendSticker(stickerId):
-    "Returns stickerSender"
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendSticker(chatId, stickerId)
-
-    return action
-
-def sendLocation(*latlon):
-    "Returns location sender"
-
-    def action(bot, chatId, data):
-        # pylint: disable=unused-argument,missing-docstring
-        bot.sendLocation(chatId, *latlon)
-
-    return action
-
-class Response():
-    "Response class"
-
-    __alowedRules = {
-        'sendMessages': sendMessages,
-        'sendMessage': sendMessage,
-        'sendKeyboard': sendKeyboard,
-        'sendSticker': sendSticker,
-        'sendPhoto': sendPhoto,
-        'sendLocation': sendLocation
-    }
-
-    def __addAction(self, actionName, actionArgs):
-        if isinstance(actionArgs, list):
-            action = self.__alowedRules[actionName](*actionArgs)
-        else:
-            action = self.__alowedRules[actionName](actionArgs)
-        self.actions.append(action)
-
-    def __processResponseDescription(self, response):
-        for key in response:
-            if key in self.__alowedRules:
-                self.__addAction(key, response[key])
-            else:
-                raise RuleNameError('Response property: "{}" is invalid'.format(key))
-
-    def __init__(self, responseBody):
-        self.body = responseBody
-        self.actions = []
-
-        responses = flatten(processResponseBody(responseBody))
-
-        for response in responses:
-            self.__processResponseDescription(response)
-
-    def run(self, update, bot):
-        """
-            Runs response actions
-        """
         message = update.get('message', {})
         text = message.get('text')
         sender = message.get('from', {})
@@ -145,5 +56,288 @@ class Response():
             'second_name': sender.get('second_name')
         }
 
-        for action in self.actions:
-            action(bot, senderId, data)
+        methodArgs = self.getData(data)
+
+        if not instanceof(methodArgs, [dict, tuple]):
+            methodArgs = [methodArgs]
+
+        sendArguments = [senderId] + list(methodArgs)
+
+        options = self.getOptions()
+        options = {'options': options}
+
+        return getattr(bot, self.method)(*sendArguments, **options)
+
+class Text(Message):
+    "Text response container"
+    # pylint: disable=missing-docstring
+
+    method = 'sendMessage'
+
+    def __init__(self, text, **options):
+        super().__init__(**options)
+        self.text = text
+
+    def getData(self, data):
+        text = self.text
+
+        if self.options.get('interpolate'):
+            text = self.text.format(
+                text=data.get('text'),
+                first_name=data.get('first_name'),
+                username=data.get('username'),
+                second_name=data.get('second_name'))
+
+        return text
+
+    def getOptions(self):
+        options = super().getOptions()
+        print('options: {}'.format(options))
+
+        additional = {
+            'disable_web_page_preview': self.options.get('disableWebPreview', False),
+        }
+
+        options.update(additional)
+
+        if self.options.get('format'):
+            options.update({'parse_mode': self.options.get('format')})
+
+        if self.options.get('markup'):
+            options.update({'reply_markup': self.options.get('markup')})
+
+        return options
+
+class Keyboard(Text):
+    "Keyboard response container"
+    method = 'sendKeyboard'
+    # pylint: disable=missing-docstring
+
+    @staticmethod
+    def __toKeyboard(row):
+        def toButton(btn):
+            if isinstance(btn, dict) and btn['text']:
+                return btn
+            else:
+                return {
+                    'text': str(btn)
+                }
+        return list(map(toButton, row))
+
+    def __init__(self, text, keyboard=None, **options):
+        #pylint: disable=redefined-variable-type
+        super().__init__(text, **options)
+        self.text = text if keyboard else text.get('text')
+        self.keyboard = keyboard if keyboard else text.get('keyboard', [])
+
+        self.keyboard = list(map(self.__toKeyboard, self.keyboard))
+
+        if isinstance(self.keyboard, list):
+            self.keyboard = {
+                'keyboard': self.keyboard
+            }
+
+        if options.get('autohide') != None:
+            self.keyboard['one_time_keyboard'] = options.get('autohide')
+
+        if options.get('resize') != None:
+            self.keyboard['resize_keyboard'] = options.get('resize')
+
+    def getData(self, data):
+        return super().getData(data), self.keyboard
+
+class Location(Message):
+    "Location response container"
+    # pylint: disable=missing-docstring
+    method = 'sendLocation'
+
+    def __init__(self, lat, lon, **options):
+        super().__init__(**options)
+        self.lat = lat
+        self.lon = lon
+
+    def getData(self, data):
+        return self.lat, self.lon
+
+class Sticker(Message):
+    "Sticker response container"
+    # pylint: disable=missing-docstring
+    method = 'sendSticker'
+
+    def __init__(self, sticker, **options):
+        super().__init__(**options)
+        self.sticker = sticker
+
+    def getData(self, data):
+        return self.sticker
+
+
+class File(Message):
+    "File response container"
+    # pylint: disable=missing-docstring
+
+    method = 'sendDocument'
+
+    def __init__(self, file, caption=None, **options):
+        super().__init__(**options)
+        self.file = file
+        self.caption = caption
+
+    def getData(self, data):
+        return self.file, self.caption
+
+
+class Photo(File):
+    "Photo response container"
+    method = 'sendPhoto'
+
+class Document(File):
+    "Document response container"
+    method = 'sendDocument'
+
+class Voice(File):
+    "Voice response container"
+    method = 'sendVoice'
+
+class Audio(File):
+    "Audio response container"
+    method = 'sendAudio'
+
+class Video(File):
+    "Video response container"
+    method = 'sendVideo'
+
+class Contact(Message):
+    "Contact response container"
+    # pylint: disable=missing-docstring
+    method = 'sendContact'
+
+    def __init__(self, phone, first, last='', **options):
+        super().__init__(**options)
+        self.phone = phone
+        self.first = first
+        self.last = last
+
+    def getData(self, data):
+        return self.phone, self.first, self.last
+
+def getAction(actionType, data):
+    "Calculates action by actionType in response description"
+    # pylint: disable=missing-docstring,unnecessary-lambda,redefined-variable-type,too-many-locals,too-many-branches
+    response = None
+
+    if actionType == 'text':
+        if isinstance(data, dict):
+            text = data.get('text', '')
+            data = omit(data, ['text'])
+            response = Text(text, **data)
+        else:
+            text = data
+            response = Text(text)
+    elif actionType == 'location':
+        if isinstance(data, dict):
+            lat = data.get('lat')
+            lon = data.get('lon')
+            options = omit(data, ['lat', 'lon'])
+            response = Location(lat, lon, **options)
+        elif isinstance(data, list):
+            response = Location(data[0], data[1])
+        else:
+            raise ResponseFormatError('Invalid format at Location response')
+    elif actionType == 'keyboard':
+        if isinstance(data, dict):
+            optionNames = ['autohide', 'resize']
+            options = pickCompat(data, optionNames)
+            keyboard = omit(data, optionNames)
+            response = Keyboard(keyboard, **options)
+        else:
+            raise ResponseFormatError('Invalid format at Keyboard response')
+    elif actionType == 'photo':
+        if isinstance(data, dict):
+            photo = data.get('photo')
+            caption = data.get('caption')
+            options = omit(data, ['photo', 'caption'])
+            response = Photo(photo, caption, **options)
+        else:
+            response = Photo(data)
+    elif actionType == 'sticker':
+        if isinstance(data, dict):
+            sticker = data.get('sticker')
+            options = omit(data, ['sticker'])
+            response = Sticker(sticker, **options)
+        else:
+            response = Sticker(data)
+    elif actionType == 'contact':
+        if isinstance(data, dict):
+            phone = data.get('phone')
+            name = data.get('firstName')
+            lastName = data.get('lastName')
+            options = omit(data, ['phone', 'firstName', 'lastName'])
+            response = Contact(phone, name, lastName, **options)
+        else:
+            raise ResponseFormatError('Invalid format at Contact response')
+    elif actionType == 'voice' or actionType == 'audio':
+        if isinstance(data, dict):
+            file = data.get(actionType)
+            caption = data.get('caption')
+            options = omit(data, ['voice', 'caption', 'audio'])
+            response = Voice(file, caption, **options) if actionType == 'voice' else Audio(file, caption, **options)
+        else:
+            typeOfResponse = 'Voice' if actionType == 'voice' else 'Audio'
+            raise ResponseFormatError('Invalid format at {} response'.format(typeOfResponse))
+    elif actionType == 'video':
+        if isinstance(data, dict):
+            file = data.get('video')
+            caption = data.get('caption')
+            options = omit(data, ['caption', 'video'])
+            response = Video(file, caption, **options)
+        else:
+            raise ResponseFormatError('Invalid format at Video response')
+    else:
+        raise ResponseMessageError('Invalid response type: "{}"'.format(actionType))
+
+    def action(response):
+        return lambda bot, update: response.run(bot, update)
+
+    return action(response)
+
+class Response():
+    "Response class"
+
+    __map = {
+        'text': Text,
+        'keyboard': Keyboard,
+        'location': Location,
+        'photo': Photo,
+        'sticker': Sticker,
+        'audio': Audio,
+        'voice': Voice
+    }
+
+    def __init__(self, responses):
+        self.body = responses
+        self.actions = []
+
+        if not isinstance(responses, list):
+            responses = [responses]
+
+        for response in responses:
+            self.addAction(response)
+
+    def addAction(self, response):
+        "Adds action to actions list"
+        if len(response) > 1:
+            raise ResponseFormatError('Response can hold only one action')
+
+        responseKey = list(response.keys()).pop()
+        respBody = response[responseKey]
+
+        if responseKey in self.__map:
+            self.action = getAction(responseKey, respBody)
+        else:
+            raise ResponseFormatError('Invalid response action: "{}"'.format(responseKey))
+
+    def run(self, bot, update):
+        "Runs response actions"
+        return self.action(bot, update)
+