+++ bobot/utils/utils.py
@@ -44,3 +44,34 @@ def getFile(f):
         @return file
     """
     return open(f, 'rb')
+
+def instanceof(obj, instances):
+    "Checks does object instance of some of given instances"
+    lst = filter(lambda instance: isinstance(obj, instance), instances)
+    return bool(len(list(lst)))
+
+def omit(obj, keys):
+    """
+        Remove keys from dict by names
+        @param {dict} obj
+        @param {list[str]} keys
+        @return {dict}
+
+    """
+    obj = obj.copy()
+    for key in keys:
+        if obj.get(key):
+            del obj[key]
+    return obj
+
+def pickCompat(obj, keys):
+    """
+        Forming dict from given one and list of keys
+        @param {dict} obj
+        @param {list[str]} keys
+        @return {dict}
+    """
+    result = {}
+    for key in keys:
+        result[key] = obj.get(key)
+    return result